requirements.txt
flask
gunicorn
torchvision==0.16.0
numpy
pillow
pillow-heif